requirements.txt
blinker==1.8.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.4.0
reportlab==4.2.2
Werkzeug==3.0.4
WTForms==3.1.2
